requirements.txt
# Core FastAPI dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic[email]
pydantic-settings==2.3.3

# Database
asyncpg==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
cryptography==42.0.8

# Email
resend==2.2.0
jinja2==3.1.4

# AWS
boto3==1.34.131

# Redis & Caching
redis==5.0.6
aioredis==2.0.1

# Validation & Utilities
email-validator==2.1.2
pytz==2024.1
python-dateutil==2.9.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0

# Development
autopep8==2.2.0
pylint==3.2.3

# Rate limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.5.1

# Background tasks
celery==5.4.0
flower==2.0.1